Rename MergeResplitter to Merger
